refactor: port should be u16 instead of u32

+++ src/meta/service/src/meta_service/meta_node.rs
@@ -318,7 +318,7 @@ impl MetaNode {
     pub async fn start_grpc(
         mn: Arc<MetaNode>,
         host: &str,
-        port: u32,
+        port: u16,
     ) -> Result<(), MetaNetworkError> {
         let mut rx = mn.running_rx.clone();
